QUE dIOS ME AYUDE

+++ src/main/java/com/final_project/descubri_cba/model/Destination.java
@@ -53,7 +53,7 @@ public abstract class Destination {
     @OneToMany(mappedBy = "destination", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Comment> comments = new ArrayList<>();
 
-    private int averageScore;
+    private Double averageScore;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
